requirement.txt
pandas==2.2.3
prefect==3.0.10
pyarrow==17.0.0
google-cloud-bigquery==3.26.0

